dividindo responsabilidade
